refactor(api): refactor the route to get user informations
