--- api/src/shared/routes/user-route.js
@@ -1,10 +1,0 @@
-import express from "express";
-
-import userController from "../controllers/user-controller.js";
-import { checkUserToken } from "../middlewares/check-user-token.js";
-
-const userRoute = express.Router();
-
-userRoute.get("/api/authenticated", checkUserToken, userController);
-
-export default userRoute;
